Refactor Cypress configuration to use external task functions
